# rms-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RMS API Documentation",
        "description": "API du Revenu Minimum Solidaire (RMS) : auth, portefeuille, file FIFO, rotations, timeline.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/queue/position": {
            "get": {
                "tags": [
                    "Queue"
                ],
                "summary": "Obtenir la position d'un membre dans la file FIFO",
                "operationId": "b81da9c32e0467e0d0244d82957019d1",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Position retournée"
                    },
                    "404": {
                        "description": "Utilisateur ou entrée de file introuvable"
                    }
                }
            }
        },
        "/api/admin/rotations/run-once": {
            "post": {
                "tags": [
                    "Admin - Rotations"
                ],
                "summary": "Déclenche une rotation RMS unique",
                "description": "Exécute une seule rotation : sélectionne un membre éligible (remboursé-éligible ou premier de la file FIFO), crédite son gain de 100 000 FCFA et met à jour les historiques.",
                "operationId": "adminRunSingleRotation",
                "responses": {
                    "201": {
                        "description": "Rotation exécutée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "rotation_id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "amount": {
                                                    "type": "integer"
                                                },
                                                "source": {
                                                    "description": "Source des fonds (ex: jackpot)",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Rotation exécutée avec succès.",
                                    "data": {
                                        "rotation_id": 1,
                                        "user_id": 3,
                                        "amount": 100000,
                                        "source": "jackpot"
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Aucune rotation possible pour le momKent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Aucune rotation possible pour le moment (cagnotte ou candidats insuffisants)."
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}